ECEID implemented
ECEID and tests implemented, ported from code written in June 2020. The energy is conserver to numerical precision basically, this is not the case for Ehrenfest where the energy conservation is of the order of the time step. There is a odd thing in the plots of the expectation value of sigma z, the path consists of straight lines joining three time steps, this is an odd feature of the integrator or I am bootstrapping it in the wrong way.