"Constant" matrices replaced by functions
There is no such thing as a "constant" matrix (!!!), found out the hard way (ie: modifying one by reference when modifying a different one that was made "=" to the constant. Therefore, all constant matrices have been replaced by functions. This increases the overhead allocation by a crazy amount, doubling the time for the CEID case in the tests since every call to H, F or K requires to allocate extra space.

One could use them if instead of = one uses .= (I think) however, this is no good when passing them as parameter to functions that are then used as initializers.